Fixed inputs and outputs
In the plugins XML file, the InputProperty element's name
attribute *must* be named "Input". If not, the filter will
not appear correctly as the child of its input in the
pipeline browser. For two-input filters, the name can be
anything, so I named the first two FirstInput and SecondInput.

Also, used a DeepCopy() for the output rather than a ShallowCopy()
because the original data must be garbage collected within
GenerateData().